requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.3.0
